# TP03/Amigos.API/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Amigos.API",
    "version": "v1"
  },
  "paths": {
    "/api/amigos": {
      "post": {
        "tags": [
          "Amigos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CadastrarAmigoRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Amigos"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Amigos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CadastrarAmigoRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Amigos"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "CadastrarAmigoRequest": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string"
        },
        "sobrenome": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "telefone": {
          "format": "int32",
          "type": "integer"
        },
        "aniversario": {
          "format": "date-time",
          "type": "string"
        }
      }
    }
  }
}